Bug 1844020 - Rework PipeWire camera enumeration r=pehrsons
1) Avoid using V4L2 backend when PipeWire is enabled
When WebRTC backend was initialized without using VideoCaptureOptions it
always defaulted to use V4L2 backend as PipeWire can be only used when
options are passed with "allow_pipewire=true". We now use options right
from the beginning, setting allow_pipewire=true" or "allow_v4l2=true"
depending on the backend we intend to use.

2) Use a camera placeholder when PipeWire is not fully initialized
In scenario where websites do an enumeration without being able to ask
for permissions, we would not have access to PipeWire socket and not
being able to list camera devices. In this case we return a placeholder
to indicate there is a camera device and later when there is a proper
gUM request, we will be able to ask for permissions and enable access
to PipeWire to give back the full list of camera devices.

Differential Revision: https://phabricator.services.mozilla.com/D188217