Bug 1844020 - Use VideoCaptureOptions to initialize VideoCapture and DeviceInfo also for V4L2 r=pehrsons
With the introduction of PipeWire into WebRTC, there have been new APIs
added, specifically a new way how VideoCapture and DeviceInfo can be
created by passing VideoCaptureOptions. This has been the only way how
to use PipeWire backend, while not passing VideoCaptureOptions we always
defaulted to V4L2. Use this new API for both directly so we avoid using
V4L2 when we intend to use PipeWire, which we originally were only been
able to use later once we were granted permissions to PipeWire socket.

This change introduces VideoCaptureFactory handling VideoCaptureOptions
initialization and creation of DeviceInfo and VideoCaptureModule that is
now moved from VideoEngine. The new VideoCaptureFactory takes care of
platform and backend specific code, leaving VideoEngine to be as generic
as possible.

Differential Revision: https://phabricator.services.mozilla.com/D188217